refactor: extract tables to organisms. Also review comments
