+++ web-registry/src/components/atoms/BreakText.tsx
@@ -1,0 +1,6 @@
+import { styled } from '@mui/material';
+
+export const BreakText = styled('div')({
+  whiteSpace: 'normal',
+  wordWrap: 'break-word',
+});
