+++ web-registry/src/components/atoms/BreakTextEnd.tsx
@@ -1,0 +1,7 @@
+import { styled } from '@mui/material';
+
+export const BreakTextEnd = styled('div')({
+  whiteSpace: 'normal',
+  wordWrap: 'break-word',
+  textAlign: 'end',
+});
