+++ web-registry/src/components/atoms/GreyText.tsx
@@ -1,0 +1,5 @@
+import { styled } from '@mui/material';
+
+export const GreyText = styled('span')(({ theme }) => ({
+  color: theme.palette.info.main,
+}));
