+++ web-registry/src/components/organisms/BridgableEcocreditsTable/BridgableEcocreditsTable.constants.ts
@@ -3,9 +3,4 @@ export const CREDIT_BATCH_TOOLTIP =
   'Each credit batch has a unique ID (i.e. denomination) that starts with the abbreviation of the credit type followed by the project ID, start date and end date of the vintage year, and ending in the batch sequence number.';
 export const AMOUNT_BRIDGABLE_TOOLTIP =
   'The amount bridgable is the same as the amount tradable. Credits that have been retired cannot be bridged.';
-export const AMOUNT_BRIDGED_TOOLTIP =
-  'Amount bridged is the same as amount cancelled in the ledger documentation';
 export const NO_BRIDGABLE_CREDITS = 'No bridgable ecocredits found';
-export const NO_BRIDGED_CREDITS = 'No bridged ecocredits found';
-
-export const BRIDGED_STATUS = ['pending', 'complete', 'error'];
