+++ web-registry/src/components/organisms/BridgableEcocreditsTable/BridgableEcocreditsTable.styles.ts
@@ -1,0 +1,8 @@
+export const withLoaderStyles = {
+  display: 'flex',
+  width: '100%',
+  justifyContent: 'center',
+  flexDirection: 'col',
+  alignItems: 'center',
+  py: 25,
+};
