+++ web-registry/src/components/organisms/BridgedEcocreditsTable/BridgedEcocreditsTable.constants.ts
@@ -1,0 +1,6 @@
+export const BRIDGED_STATUS = ['pending', 'complete', 'error'];
+export const CREDIT_BATCH_TOOLTIP =
+  'Each credit batch has a unique ID (i.e. denomination) that starts with the abbreviation of the credit type followed by the project ID, start date and end date of the vintage year, and ending in the batch sequence number.';
+export const NO_BRIDGED_CREDITS = 'No bridged ecocredits found';
+export const AMOUNT_BRIDGED_TOOLTIP =
+  'Amount bridged is the same as amount cancelled in the ledger documentation';
