+++ web-registry/src/components/organisms/BridgedEcocreditsTable/BridgedEcocreditsTable.styles.ts
@@ -1,0 +1,10 @@
+export const withLoaderStyles = {
+  display: 'flex',
+  width: '100%',
+  justifyContent: 'center',
+  flexDirection: 'col',
+  alignItems: 'center',
+  py: 25,
+};
+
+export const iconStyle = { height: '16px', verticalAlign: 'middle', mb: '2px' };
