+++ web-registry/src/pages/Dashboard/MyBridge/MyBridge.tsx
@@ -1,18 +1,31 @@
 import { Box } from 'web-components/lib/components/box';
+import Card from 'web-components/lib/components/cards/Card';
+import { IconTabProps } from 'web-components/lib/components/tabs/IconTab';
+import { IconTabs } from 'web-components/lib/components/tabs/IconTabs';
 
-import { BridgeTabs } from './BridgeTabs';
+import { BridgableEcocreditsTable } from 'components/organisms/BridgableEcocreditsTable/BridgableEcocreditsTable';
+import { BridgedEcocreditsTable } from 'components/organisms/BridgedEcocreditsTable/BridgedEcocreditsTable';
+
 import { BridgeInfo } from './MyBridge.BridgeInfo';
+import { cardStyles, tabsStyles, wrapperStyles } from './MyBridge.styles';
+
+const tabs: IconTabProps[] = [
+  {
+    label: 'Bridgable ecocredits',
+    content: <BridgableEcocreditsTable />,
+  },
+  {
+    label: 'Bridged ecocredits',
+    content: <BridgedEcocreditsTable />,
+  },
+];
 
 export const MyBridge = (): JSX.Element => {
   return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        width: '100%',
-      }}
-    >
-      <BridgeTabs />
+    <Box sx={wrapperStyles}>
+      <Card sx={cardStyles}>
+        <IconTabs tabs={tabs} size={'xl'} sxs={tabsStyles} />
+      </Card>
       <BridgeInfo />
     </Box>
   );
